--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@
 /target
 /.idea
 /*.iml
+*.jfr