--- a/src/main/java/com/arangodb/velocypack/VPackBuilder.java
+++ b/src/main/java/com/arangodb/velocypack/VPackBuilder.java
@@ -216,7 +216,7 @@ public class VPackBuilder {
 	private int size;
 	private final List<Integer> stack; // Start positions of open
 										// objects/arrays
-	private final Map<Integer, List<Integer>> index; // Indices for starts
+	private List<Integer>[] index; // Indices for starts
 														// of
 														// subindex
 	private boolean keyWritten; // indicates that in the current object the key
@@ -233,7 +233,7 @@ public class VPackBuilder {
 		size = 0;
 		buffer = new byte[10];
 		stack = new ArrayList<Integer>();
-		index = new HashMap<Integer, List<Integer>>();
+		index = new List[4];
 	}
 
 	public BuilderOptions getOptions() {
@@ -670,7 +670,11 @@ public class VPackBuilder {
 	private void addCompoundValue(final byte head) {
 		// an Array or Object is started:
 		stack.add(size);
-		index.put(stack.size() - 1, new ArrayList<Integer>());
+		if (index.length < stack.size()) {
+			// ensureCapacity
+			index = Arrays.copyOf(index, index.length * 2);
+		}
+		index[stack.size() - 1] = new ArrayList<Integer>();
 		add(head);
 		// Will be filled later with bytelength and nr subs
 		size += 8;
@@ -682,12 +686,12 @@ public class VPackBuilder {
 	}
 
 	private void reportAdd() {
-		final Collection<Integer> depth = index.get(stack.size() - 1);
+		final Collection<Integer> depth = index[stack.size() - 1];
 		depth.add(size - stack.get(stack.size() - 1));
 	}
 
 	private void cleanupAdd() {
-		final List<Integer> depth = index.get(stack.size() - 1);
+		final List<Integer> depth = index[stack.size() - 1];
 		depth.remove(depth.size() - 1);
 	}
 
@@ -708,7 +712,7 @@ public class VPackBuilder {
 		}
 		final byte head = head();
 		final boolean isArray = head == 0x06 || head == 0x13;
-		final List<Integer> in = index.get(stack.size() - 1);
+		final List<Integer> in = index[stack.size() - 1];
 		final int tos = stack.get(stack.size() - 1);
 		if (in.isEmpty()) {
 			return closeEmptyArrayOrObject(tos, isArray);
@@ -990,10 +994,10 @@ public class VPackBuilder {
 	}
 
 	private static class SortEntry {
-		private final VPackSlice slice;
+		private final VPackStringSlice slice;
 		private final int offset;
 
-		public SortEntry(final VPackSlice slice, final int offset) {
+		public SortEntry(final VPackStringSlice slice, final int offset) {
 			super();
 			this.slice = slice;
 			this.offset = offset;
@@ -1002,14 +1006,14 @@ public class VPackBuilder {
 
 	private void sortObjectIndex(final int start, final List<Integer> offsets)
 			throws VPackKeyTypeException, VPackNeedAttributeTranslatorException {
-		final List<VPackBuilder.SortEntry> attributes = new ArrayList<VPackBuilder.SortEntry>();
+		final List<VPackBuilder.SortEntry> attributes = new ArrayList<VPackBuilder.SortEntry>(offsets.size());
 		for (final Integer offset : offsets) {
-			attributes.add(new SortEntry(new VPackSlice(buffer, start + offset).makeKey(), offset));
+			attributes.add(new SortEntry(new VPackSlice(buffer, start + offset).makeKey().getAsStringSlice(), offset));
 		}
 		final Comparator<SortEntry> comparator = new Comparator<SortEntry>() {
 			@Override
 			public int compare(final SortEntry o1, final SortEntry o2) {
-				return o1.slice.getAsString().compareTo(o2.slice.getAsString());
+				return o1.slice.compareTo(o2.slice);
 			}
 		};
 		Collections.sort(attributes, comparator);