--- a/src/main/java/com/arangodb/velocypack/internal/util/ObjectArrayUtil.java
+++ b/src/main/java/com/arangodb/velocypack/internal/util/ObjectArrayUtil.java
@@ -29,66 +29,59 @@ import java.util.Map;
  */
 public class ObjectArrayUtil {
 
-	private static final Map<Byte, Integer> FIRST_SUB_MAP;
-
-	static {
-		FIRST_SUB_MAP = new HashMap<Byte, Integer>();
-
-		FIRST_SUB_MAP.put((byte) 0x00, 0); // None
-		FIRST_SUB_MAP.put((byte) 0x01, 1); // empty array
-		FIRST_SUB_MAP.put((byte) 0x02, 2); // array without index table
-		FIRST_SUB_MAP.put((byte) 0x03, 3); // array without index table
-		FIRST_SUB_MAP.put((byte) 0x04, 5); // array without index table
-		FIRST_SUB_MAP.put((byte) 0x05, 9); // array without index table
-		FIRST_SUB_MAP.put((byte) 0x06, 3); // array with index table
-		FIRST_SUB_MAP.put((byte) 0x07, 5); // array with index table
-		FIRST_SUB_MAP.put((byte) 0x08, 9); // array with index table
-		FIRST_SUB_MAP.put((byte) 0x09, 9); // array with index table
-		FIRST_SUB_MAP.put((byte) 0x0a, 1); // empty object
-		FIRST_SUB_MAP.put((byte) 0x0b, 3); // object with sorted index table
-		FIRST_SUB_MAP.put((byte) 0x0c, 5); // object with sorted index table
-		FIRST_SUB_MAP.put((byte) 0x0d, 9); // object with sorted index table
-		FIRST_SUB_MAP.put((byte) 0x0e, 9); // object with sorted index table
-		FIRST_SUB_MAP.put((byte) 0x0f, 3); // object with unsorted index table
-		FIRST_SUB_MAP.put((byte) 0x10, 5); // object with unsorted index table
-		FIRST_SUB_MAP.put((byte) 0x11, 9); // object with unsorted index table
-		FIRST_SUB_MAP.put((byte) 0x12, 9); // object with unsorted index table
-	}
+	private static final int[] FIRST_SUB_MAP = {
+		0, // 0x00 None
+		1, // 0x01 empty array
+		2, // 0x02 array without index table
+		3, // 0x03 array without index table
+		5, // 0x04 array without index table
+		9, // 0x05 array without index table
+		3, // 0x06 array with index table
+		5, // 0x07 array with index table
+		9, // 0x08 array with index table
+		9, // 0x09 array with index table
+		1, // 0x0a empty object
+		3, // 0x0b object with sorted index table
+		5, // 0x0c object with sorted index table
+		9, // 0x0d object with sorted index table
+		9, // 0x0e object with sorted index table
+		3, // 0x0f object with unsorted index table
+		5, // 0x10 object with unsorted index table
+		9, // 0x11 object with unsorted index table
+		9 //  0x12 object with unsorted index table
+	};
 
 	public static int getFirstSubMap(final byte key) {
-		return FIRST_SUB_MAP.get(key);
+		return FIRST_SUB_MAP[key & 0xff];
 	}
 
-	private static final Map<Byte, Integer> OFFSET_SIZE;
-	static {
-		OFFSET_SIZE = new HashMap<Byte, Integer>();
-
-		OFFSET_SIZE.put((byte) 0x00, 0); // None
-		OFFSET_SIZE.put((byte) 0x01, 1); // empty array
-		OFFSET_SIZE.put((byte) 0x02, 1); // array without index table
-		OFFSET_SIZE.put((byte) 0x03, 2); // array without index table
-		OFFSET_SIZE.put((byte) 0x04, 4); // array without index table
-		OFFSET_SIZE.put((byte) 0x05, 8); // array without index table
-		OFFSET_SIZE.put((byte) 0x06, 1); // array with index table
-		OFFSET_SIZE.put((byte) 0x07, 2); // array with index table
-		OFFSET_SIZE.put((byte) 0x08, 4); // array with index table
-		OFFSET_SIZE.put((byte) 0x09, 8); // array with index table
-		OFFSET_SIZE.put((byte) 0x0a, 1); // empty object
-		OFFSET_SIZE.put((byte) 0x0b, 1); // object with sorted index table
-		OFFSET_SIZE.put((byte) 0x0c, 2); // object with sorted index table
-		OFFSET_SIZE.put((byte) 0x0d, 4); // object with sorted index table
-		OFFSET_SIZE.put((byte) 0x0e, 8); // object with sorted index table
-		OFFSET_SIZE.put((byte) 0x0f, 1); // object with unsorted index table
-		OFFSET_SIZE.put((byte) 0x10, 2); // object with unsorted index table
-		OFFSET_SIZE.put((byte) 0x11, 4); // object with unsorted index table
-		OFFSET_SIZE.put((byte) 0x12, 8); // object with unsorted index table
-	}
+	private static final int[] OFFSET_SIZE = {
+		0, // 0x00 None
+		1, // 0x01 empty array
+		1, // 0x02 array without index table
+		2, // 0x03 array without index table
+		4, // 0x04 array without index table
+		8, // 0x05 array without index table
+		1, // 0x06 array with index table
+		2, // 0x07 array with index table
+		4, // 0x08 array with index table
+		8, // 0x09 array with index table
+		1, // 0x0a empty object
+		1, // 0x0b object with sorted index table
+		2, // 0x0c object with sorted index table
+		4, // 0x0d object with sorted index table
+		8, // 0x0e object with sorted index table
+		1, // 0x0f object with unsorted index table
+		2, // 0x10 object with unsorted index table
+		4, // 0x11 object with unsorted index table
+		8 //  0x12 object with unsorted index table
+	};
 
 	private ObjectArrayUtil() {
 		super();
 	}
 
 	public static int getOffsetSize(final byte key) {
-		return OFFSET_SIZE.get(key);
+		return OFFSET_SIZE[key & 0xff];
 	}
 }