… pom.xml
… src/main/java/com/arangodb/velocypack/internal/util/ValueLengthUtil.java
… src/main/java/com/arangodb/velocypack/internal/util/ValueTypeUtil.java
… src/test/resources/api-docs.json