# Summary
Adds simple benchmarks for `VPackBuilder`, `VPackSlice`, and `VPackParser`. Improves runtime performance by ~110% within `VPackBuilder`/`VPackSlice`. `String` => `VPackSlice` is dominated by json parsing and remains about the same.

## Results
https://jmh.morethan.io/?sources=https://gist.githubusercontent.com/htmldoug/d292b301fdc3ec66cd7098e4ee54e965/raw/14e54eb2c9b153ba81931d16b276511eb3796f7f/jmh-result-before.json,https://gist.githubusercontent.com/htmldoug/d292b301fdc3ec66cd7098e4ee54e965/raw/14e54eb2c9b153ba81931d16b276511eb3796f7f/jmh-result-optimized.json

## Context
I don't use arangodb, but I'm interested in using velocypack to reduce the heap usage of my json processing. I recently saw a 35 MB byte array get parsed to 270 MB of heap by play-json's `JsValue`. [Case classes ate my RAM](https://www.youtube.com/watch?v=OtBTBaA5M4g) inspired backing the `JsValue` by a `VPackSlice` and it drops the heap usage to ~38 MB for the cost of about 33% more CPU time.

JMH + JFR + JMC revealed some some low-hanging fruit to cut the extra CPU time roughly in half for my use case. I also benched a real world example of String => VPackSlice [from arrangodb/velocypack](https://github.com/arangodb/velocypack/blob/master/tests/jsonSample/api-docs.json). The majority of time is spent on json parsing, although throughput still improved by 20%.

## Optimizations
The gains come mostly from reducing UTF-8 encoding by `string.getBytes()` and replacing `HashMap<Integer, T>` lookups for sequential indexes with simple `T[]` arrays.

## CLA
I've emailed a signed CLA to cla@arangodb.com.