@siilike 2019-11-15
It is sad to see that such great (and obvious!) performance improvements take so long to get accepted.