@rashtao 2019-11-19
@htmldoug I am reviewing the PR, but there are some points unclear to me:
1. the code has many warnings, eg. unused imports and Javadoc warnings 
2. the code is not compatible with java 1.6, which is currently the base java version of the project
3. using `-XX:+UnlockCommercialFeatures` and `JmhFlightRecorderProfiler` is not compatible with our license and cannot be used with opensource software, at least until we will use jdk 11
4. running the benchmark I see no improvement for `Bench.fromJson`, actually it gets slightly worse (0.3% slower). In the other 2 benchmarks I get results similar to yours.

I would respectively suggest you to: 
1. fix 
2. set the minimum required java version to `1.7` for the project
3. remove `-XX:+UnlockCommercialFeatures` and `JmhFlightRecorderProfiler`
4. you can leave `Bench.fromJson` as is, so it will help us in the future

Also, upgrading to java 7 as lower required java version, would enable us to use ByteBuffers, thus avoiding `Arrays.copy` every time that we have to resize the buffers, thus achieving imho considerable performance improvements. 

Thanks for contributing and sorry for the late reply!