@htmldoug 2020-01-08
> Also, upgrading to java 7 as lower required java version, would enable us to use ByteBuffers, thus avoiding `Arrays.copy` every time that we have to resize the buffers, thus achieving imho considerable performance improvements.

Hey @rashtao, how would this work? I only have a light understanding of `ByteBuffer`s, but my impression was that `HeapByteBuffer` and `DirectByteBuffer` were fixed capacity, i.e. not resizable. Are you picturing composing these into a composite like akka's `ByteString`? How were you envisioning this working?