@rashtao 2020-01-09
@htmldoug 
I mean `DirectByteBuffer` would have higher write throughput than `byte[]` and not impact on GC. I have no experience with akka's `ByteString`, but I think also composing could lead to additional performance improvements, eg. something like Netty's `CompositeByteBuf`.