@htmldoug 2020-01-09
`ensureCapacity()` is producing a ton of non-TLAB allocations. Growing the array by 1.5x rather than the more typical 2x is causing a lot of churn.

I'm still not entirely sure how `DirectByteBuffer` would help, but a composite (at least within the builder) would be really good here. I started exploring that direction, but one barrier to implementing is that `VPackBuilder.remove()` is doing deletions via some left-shifting of the array that I haven't quite figured out yet and I ran out of time.