@htmldoug 2020-01-11
@siilike, yeah, that'd be an improvement and there's good [precedent for that](https://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/util/FastByteArrayOutputStream.html).

Dreaming for a moment, I'd love the option to be able to back both `VPackBuilder` and `VPackSlice` entirely on disk for situations where I'm dealing with hundreds-of-megabytes json files that my apps sometimes receive.

This would require decoupling `VPackBuilder`/`Slice` from their backing implementation, which could be done by identifying all the operations currently used (`appendByte/s`(...), `readByte/s(...)`, even `leftShift()`?), and stuffing them into some pluggable `interface VPackStore {...}`. That'd enable your `ListByteVPackStore`, the current baseline `CompactArrayByteVPackStore`, @rashtao's `DirectByteBufferVPackStore`, or a kafka-style `RandomAccessFileVPackStore` for my stupidly large inputs.